+++ VicePenguins/src/layouts/DefaultLayout.vue
@@ -9,8 +9,8 @@
   </template>
   
   <script setup>
-  import Navbar from '../components/Navbar.vue';
-  import Footer from '../components/Footer.vue';
+  import Navbar from '../components/Navbar/Navbar.vue';
+  import Footer from '../components/Footer/Footer.vue';
   </script>
   
   <style>
